+++ containers/ecr-viewer/src/app/api/services/dialects/sqlserver.ts
@@ -1,6 +1,9 @@
-import { MssqlDialect } from "kysely";
+import { Kysely, MssqlDialect } from "kysely";
 import * as tarn from "tarn";
 import * as tedious from "tedious";
+
+import { Core } from "@/app/api/services/types/core";
+import { Extended } from "@/app/api/services/types/extended";
 
 export const dialect = {
   dialect: new MssqlDialect({
@@ -41,4 +44,19 @@ export const dialect = {
       },
     },
   }),
+};
+
+/**
+ * construct a sql server db instance
+ * @param schema core or extended
+ * @returns sql server db instance
+ */
+export const sqlServerConstructor = (schema: "core" | "extended") => {
+  if (schema === "core") {
+    return new Kysely<Core>(dialect);
+  } else if (schema === "extended") {
+    return new Kysely<Extended>(dialect);
+  } else {
+    throw new Error("Invalid schema type.");
+  }
 };
